refactor: streamline email normalization and validation across contact-related routes
